tests: merge msteams thread authz coverage
